Задача5 дубль4
